Move list for triple column demo to separate file
